split next image import to be per game
