+++ modules/next/commands/importGameImages.js
@@ -1,20 +1,14 @@
-const path = require("path");
-const fs = require("fs-extra");
 const list = require("../../games/dist/gameList");
+const importGameImages = require("./helpers/importGameImages");
 
-console.log("---------- importing all game content images ---------- ");
-for (const gameId of list) {
-  const source = path.join(
-    __dirname,
-    `../../content/material/games/${gameId}/pics`
-  );
-  if (fs.existsSync(source)) {
-    const target = path.join(__dirname, `../public/images/${gameId}`);
-    fs.removeSync(target);
-    fs.copySync(source, target);
-    console.log("images imported for", gameId);
-  } else {
-    console.log("No images found for", gameId);
+const gameId = process.argv[2];
+
+if (!gameId) {
+  console.log("---------- importing all game content images ---------- ");
+  for (const gameId of list) {
+    importGameImages(gameId);
   }
+  console.log("------- all game content images imported---------- ");
+} else {
+  importGameImages(gameId);
 }
-console.log("------- all game content images imported---------- ");
